// web/src/gen/model/contentItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JsonDocument } from "./jsonDocument";

export interface ContentItemDto {
  id: string;
  contentTypeId?: string;
  projectId?: string;
  name?: string;
  resolvedData?: JsonDocument;
  values?: JsonDocument;
  /** @nullable */
  thumbnailAssetId?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
}